Updated movie list logic.

+++ network/src/main/java/app/movies/network/model/PageDto.kt
@@ -3,7 +3,7 @@ package app.movies.network.model
 import com.google.gson.annotations.SerializedName
 
 data class PageDto(
-    @SerializedName("page") val page: Long,
+    @SerializedName("page") val page: Int,
     @SerializedName("results") val results: List<MovieDto>,
     @SerializedName("total_pages") val totalPages: Long,
     @SerializedName("total_results") val totalResults: Long,
